wrote code for auth and user login

+++ backend/controllers/sauces.js
@@ -71,15 +71,27 @@ exports.modifySauce = (req, res, next) => {
 };
 
 exports.deleteSauce = (req, res, next) => {
-  Sauce.deleteOne({ _id: req.params.id })
-    .then(() => {
-      res.status(200).json({
-        message: "Deleted!",
+  Sauce.findOne({ _id: req.params.id }).then((sauce) => {
+    if (!sauce) {
+      return res.status(404).json({
+        error: new Error("Objet non trouvé !"),
       });
-    })
-    .catch((error) => {
-      res.status(400).json({
-        error: error,
+    }
+    if (sauce.userId !== req.auth.userId) {
+      return res.status(401).json({
+        error: new Error("Requête non autorisée !"),
       });
-    });
+    }
+    Sauce.deleteOne({ _id: req.params.id })
+      .then(() => {
+        res.status(200).json({
+          message: "Deleted!",
+        });
+      })
+      .catch((error) => {
+        res.status(400).json({
+          error: error,
+        });
+      });
+  });
 };
